Add `configure.sh`
This commit removes all installation targets from Makefile, and puts
them in a separate `configure.sh` script.

The usage is as follows:

./configure --help - Prints usage
./configure --dev  - Install dependencies required for SOCI developments
(e.g., check tools, flatc)
./configure --install-zlib - Install the zlib library and headers. This
is kept as a separate arguments since users may already have zlib
installed on their systems in which case we do not want to override the
system default.

Signed-off-by: Rishabh Singhvi <rdpsin@amazon.com>